+++ src/engine/engineworkerscheduler.cpp
@@ -22,10 +22,6 @@ EngineWorkerScheduler::~EngineWorkerScheduler() {
     wait();
 }
 
-void EngineWorkerScheduler::bindWorker(EngineWorker* pWorker) {
-    pWorker->setScheduler(this);
-}
-
 void EngineWorkerScheduler::workerReady(EngineWorker* pWorker) {
     if (pWorker) {
         // If the write fails, we really can't do much since we should not block
