+++ src/engine/readaheadmanager.cpp
@@ -159,7 +159,7 @@ void ReadAheadManager::hintReader(double dRate, QList<Hint>& hintList) {
 
     // SoundTouch can read up to 2 chunks ahead. Always keep 2 chunks ahead in
     // cache.
-    int length_to_cache = 2*CachingReader::kSamplesPerChunk;
+    int length_to_cache = 2 * CachingReaderWorker::kSamplesPerChunk;
 
     current_position.length = length_to_cache;
     current_position.sample = in_reverse ?
